refactor : di 모듈에 annotation 폴더 추가
