+++ di/src/main/java/com/example/di/annotation/Inject.kt
@@ -1,4 +1,4 @@
-package com.example.di
+package com.example.di.annotation
 
 @Target(AnnotationTarget.PROPERTY, AnnotationTarget.CONSTRUCTOR, AnnotationTarget.VALUE_PARAMETER)
 @Retention(AnnotationRetention.RUNTIME)
