+++ di/src/main/java/com/example/di/annotation/Singleton.kt
@@ -1,4 +1,4 @@
-package com.example.di
+package com.example.di.annotation
 
 @Target(AnnotationTarget.CLASS)
 annotation class Singleton
